fixed tutorial for pdm-storage
