+++ pandapower/runpm.py
@@ -197,40 +197,38 @@ def runpm_ots(net, julia_file=None, pp_to_pm_callback=None, calculate_voltage_an
     _runpm(net, delete_buffer_file=delete_buffer_file, pm_file_path=pm_file_path, pdm_dev_mode=pdm_dev_mode)
     read_ots_results(net)
 
-def runpm_storage_opf(net, calculate_voltage_angles=True,
+def runpm_storage_opf(net, from_time_step, to_time_step, calculate_voltage_angles=True,
                       trafo_model="t", delta=1e-8, trafo3w_losses="hv", check_connectivity=True,
                       n_timesteps=24, time_elapsed=1., correct_pm_network_data=True, silence=True,
+                      pm_solver="juniper", pm_mip_solver="cbc", pm_nl_solver="ipopt",
                       pm_model="ACPPowerModel", pm_time_limits=None, pm_log_level=0,
                       opf_flow_lim="S", charge_efficiency=1., discharge_efficiency=1., 
-                      standby_loss=1e-8, p_loss=1e-8, q_loss=1e-8, pm_tol=1e-8, pdm_dev_mode=False, **kwargs):  
+                      standby_loss=1e-8, p_loss=1e-8, q_loss=1e-8, pm_tol=1e-4, pdm_dev_mode=False, 
+                      delete_buffer_file=True, pm_file_path = None, **kwargs):  
     """
-    Runs a non-linear power system optimization with storages and time series using PowerModels.jl.
+    Runs a non-linear power system optimization with storages and time series using PandaModels.jl.
     """
+    kwargs["from_time_step"] = from_time_step
+    kwargs["to_time_step"] = to_time_step
+
     ac = True if "DC" not in pm_model else False
     net._options = {}
     _add_ppc_options(net, calculate_voltage_angles=calculate_voltage_angles,
-                      trafo_model=trafo_model, check_connectivity=check_connectivity,
-                      mode="opf", switch_rx_ratio=2, init_vm_pu="flat", init_va_degree="flat",
-                      enforce_q_lims=True, recycle=dict(_is_elements=False, ppc=False, Ybus=False),
-                      voltage_depend_loads=False, delta=delta, trafo3w_losses=trafo3w_losses)
+                     trafo_model=trafo_model, check_connectivity=check_connectivity,
+                     mode="opf", switch_rx_ratio=2, init_vm_pu="flat", init_va_degree="flat",
+                     enforce_q_lims=True, recycle=dict(_is_elements=False, ppc=False, Ybus=False),
+                     voltage_depend_loads=False, delta=delta, trafo3w_losses=trafo3w_losses)
     _add_opf_options(net, trafo_loading='power', ac=ac, init="flat", numba=True,
-                      pp_to_pm_callback=add_storage_opf_settings, julia_file="run_powermodels_mn_storage",
-                      correct_pm_network_data=correct_pm_network_data, silence=silence, pm_model=pm_model, pm_time_limits=pm_time_limits,
+                      pp_to_pm_callback=add_storage_opf_settings, julia_file="run_powermodels_multi_storage",
+                      correct_pm_network_data=correct_pm_network_data, silence=silence, pm_model=pm_model,
+                      pm_solver=pm_solver, pm_time_limits=pm_time_limits,
                       pm_log_level=pm_log_level, opf_flow_lim=opf_flow_lim, pm_tol=pm_tol, pdm_dev_mode=pdm_dev_mode)
 
-    net._options["n_time_steps"] = n_timesteps
+    net._options["n_time_steps"] = to_time_step - from_time_step
     net._options["time_elapsed"] = time_elapsed
-    
-    net._options["charge_efficiency"] = charge_efficiency
-    net._options["discharge_efficiency"] = discharge_efficiency
-    
-    net._options["standby_loss"] = standby_loss
-    net._options["p_loss"] = p_loss
-    net._options["q_loss"] = q_loss
 
-    _runpm(net)
-    storage_results = read_pm_storage_results(net)
-    return storage_results
+    _runpm(net, delete_buffer_file=delete_buffer_file, pm_file_path=pm_file_path, pdm_dev_mode=pdm_dev_mode, **kwargs)
+
 
 
 def runpm_vstab(net, pp_to_pm_callback=None, calculate_voltage_angles=True,
