+++ pandapower/test/opf/cigre_timeseries_15min.json
@@ -1,0 +1,1 @@
+{"timestep":{"0":0,"1":1,"2":2,"3":3,"4":4,"5":5,"6":6,"7":7,"8":8,"9":9,"10":10,"11":11,"12":12,"13":13,"14":14,"15":15,"16":16,"17":17,"18":18,"19":19,"20":20,"21":21,"22":22,"23":23,"24":24,"25":25,"26":26,"27":27,"28":28,"29":29,"30":30,"31":31,"32":32,"33":33,"34":34,"35":35,"36":36,"37":37,"38":38,"39":39,"40":40,"41":41,"42":42,"43":43,"44":44,"45":45,"46":46,"47":47,"48":48,"49":49,"50":50,"51":51,"52":52,"53":53,"54":54,"55":55,"56":56,"57":57,"58":58,"59":59,"60":60,"61":61,"62":62,"63":63,"64":64,"65":65,"66":66,"67":67,"68":68,"69":69,"70":70,"71":71,"72":72,"73":73,"74":74,"75":75,"76":76,"77":77,"78":78,"79":79,"80":80,"81":81,"82":82,"83":83,"84":84,"85":85,"86":86,"87":87,"88":88,"89":89,"90":90,"91":91,"92":92,"93":93,"94":94,"95":95},"pv":{"0":0.0061428571,"1":0.0058772894,"2":0.0055054945,"3":0.0052930403,"4":0.0049212454,"5":0.0046190476,"6":0.0042857143,"7":0.0039285714,"8":0.0036739927,"9":0.0033919414,"10":0.003029304,"11":0.0027472527,"12":0.0025714286,"13":0.0022857143,"14":0.0019952381,"15":0.0017333333,"16":0.0014908425,"17":0.0012600733,"18":0.0010555556,"19":0.0008571429,"20":0.0006984127,"21":0.0005311355,"22":0.0003553114,"23":0.0002222222,"24":0.0001318681,"25":0.0000659341,"26":0.000010989,"27":0.0,"28":0.0000612245,"29":0.000244898,"30":0.0003673469,"31":0.0006530612,"32":0.0010340136,"33":0.0017063492,"34":0.0024365079,"35":0.0220952381,"36":0.0538095238,"37":0.1234285714,"38":0.1466285714,"39":0.3191666667,"40":0.4307142857,"41":0.5313333333,"42":0.5887142857,"43":0.6539365079,"44":0.5895714286,"45":0.5693015873,"46":0.4239333333,"47":0.3119166667,"48":0.335,"49":0.4423142857,"50":0.4685238095,"51":0.4832619048,"52":0.4542952381,"53":0.4389428571,"54":0.2424603175,"55":0.148247619,"56":0.1802619048,"57":0.1397619048,"58":0.1086380952,"59":0.1067509158,"60":0.1093150183,"61":0.0926785714,"62":0.0563988095,"63":0.0353142857,"64":0.0254761905,"65":0.0232698413,"66":0.0222142857,"67":0.0211587302,"68":0.0203650794,"69":0.0195714286,"70":0.0189761905,"71":0.0175873016,"72":0.0168333333,"73":0.0162380952,"74":0.015452381,"75":0.0148333333,"76":0.0141836735,"77":0.0138435374,"78":0.0131031746,"79":0.0125277778,"80":0.0118359788,"81":0.0113703704,"82":0.0106709957,"83":0.0100865801,"84":0.0097359307,"85":0.0090857143,"86":0.0086857143,"87":0.008152381,"88":0.0073445378,"89":0.0067955182,"90":0.0063837535,"91":0.0058347339,"92":0.0054229692,"93":0.0047037037,"94":0.0041428571,"95":0.0036666667},"wind":{"0":0.6226,"1":0.6235846154,"2":0.6250564103,"3":0.6258974359,"4":0.6273692308,"5":0.6451384615,"6":0.6851282051,"7":0.7132666667,"8":0.7194974359,"9":0.6861128205,"10":0.6508,"11":0.6307333333,"12":0.6394,"13":0.6826,"14":0.7283733333,"15":0.77724,"16":0.8079589744,"17":0.7493384615,"18":0.6996555556,"19":0.6606,"20":0.6293555556,"21":0.5971948718,"22":0.5641179487,"23":0.5404555556,"24":0.5353589744,"25":0.5531128205,"26":0.5679076923,"27":0.5868666667,"28":0.6072285714,"29":0.5883142857,"30":0.5757047619,"31":0.5522571429,"32":0.5317904762,"33":0.5513666667,"34":0.5723,"35":0.6030333333,"36":0.6208,"37":0.6304666667,"38":0.6375866667,"39":0.6457,"40":0.6276,"41":0.5918888889,"42":0.5655666667,"43":0.5334222222,"44":0.5332666667,"45":0.5890444444,"46":0.7842666667,"47":0.9233333333,"48":1.0052166667,"49":1.0118133333,"50":0.9797333333,"51":0.9418333333,"52":0.90148,"53":0.8783333333,"54":0.8383555556,"55":0.81672,"56":0.7973,"57":0.7880444444,"58":0.7884,"59":0.7939538462,"60":0.8028564103,"61":0.80865,"62":0.8126083333,"63":0.81268,"64":0.8094333333,"65":0.75,"66":0.7076666667,"67":0.6527777778,"68":0.6118888889,"69":0.571,"70":0.5403333333,"71":0.4687777778,"72":0.48895,"73":0.5566444444,"74":0.6610666667,"75":0.7616,"76":0.754,"77":0.6961333333,"78":0.582,"79":0.5136666667,"80":0.4635407407,"81":0.5363703704,"82":0.6475151515,"83":0.7433030303,"84":0.8007757576,"85":0.8503866667,"86":0.8732266667,"87":0.90368,"88":0.932172549,"89":0.9282509804,"90":0.9253098039,"91":0.9213882353,"92":0.9184470588,"93":0.886837037,"94":0.8583393939,"95":0.838369697},"residential":{"0":0.256022,"1":0.248835,"2":0.2119733333,"3":0.201825,"4":0.209164,"5":0.1986,"6":0.18905,"7":0.186924,"8":0.1870066667,"9":0.1836575,"10":0.181048,"11":0.1855666667,"12":0.1822822222,"13":0.192295,"14":0.19936,"15":0.1822875,"16":0.186546,"17":0.192088,"18":0.203122,"19":0.22803,"20":0.2446333333,"21":0.240875,"22":0.26249,"23":0.3318133333,"24":0.43635,"25":0.501545,"26":0.52612,"27":0.53715,"28":0.614585,"29":0.6310525,"30":0.6089275,"31":0.60786,"32":0.624715,"33":0.670312,"34":0.69094,"35":0.64048,"36":0.6392,"37":0.636695,"38":0.652735,"39":0.65687,"40":0.678112,"41":0.6487575,"42":0.611575,"43":0.62915,"44":0.60457,"45":0.6140533333,"46":0.702965,"47":0.726462,"48":0.7271,"49":0.7306,"50":0.71867,"51":0.6943366667,"52":0.6500266667,"53":0.632864,"54":0.5709366667,"55":0.5649275,"56":0.565415,"57":0.5238728571,"58":0.49257,"59":0.49336,"60":0.494898,"61":0.48185,"62":0.46305,"63":0.44926,"64":0.45825,"65":0.5083566667,"66":0.51753,"67":0.5749233333,"68":0.63139,"69":0.695165,"70":0.72372,"71":0.776485,"72":0.81294,"73":0.8595785714,"74":0.86886,"75":0.902426,"76":0.9147775,"77":0.9045125,"78":0.884404,"79":0.8297333333,"80":0.8209328571,"81":0.7957233333,"82":0.738732,"83":0.70009,"84":0.6890433333,"85":0.67773,"86":0.64621,"87":0.613804,"88":0.5403266667,"89":0.4990371429,"90":0.4692085714,"91":0.43566,"92":0.41087,"93":0.344715,"94":0.304026,"95":0.2863281818}}
